Move master lua scripts to master/
