+++ nginx.conf
@@ -50,7 +50,7 @@ http {
     lua_shared_dict shmlocks 100k;
 
     init_worker_by_lua '
-        local statecache = require "mesosstatecache"
+        local statecache = require "master.mesosstatecache"
         statecache.periodically_poll_mesos_state()
     ';
 
@@ -159,7 +159,7 @@ http {
 
             more_clear_input_headers Accept-Encoding;
             rewrite ^/slave/[0-9a-zA-Z-]+/.*$ $url break;
-            rewrite_by_lua_file conf/slave.lua;
+            rewrite_by_lua_file conf/master/slave.lua;
 
             proxy_set_header        Host $http_host;
             proxy_set_header        X-Real-IP $remote_addr;
@@ -182,7 +182,7 @@ http {
 
             more_clear_input_headers Accept-Encoding;
             rewrite ^/service/[0-9a-zA-Z-.]+/?.*$ /$url break;
-            rewrite_by_lua_file conf/service.lua;
+            rewrite_by_lua_file conf/master/service.lua;
 
             proxy_set_header        Host $http_host;
             proxy_set_header        X-Real-IP $remote_addr;
@@ -205,7 +205,7 @@ http {
 
         location /metadata {
             access_by_lua 'auth.validate_jwt_or_exit()';
-            content_by_lua_file conf/metadata.lua;
+            content_by_lua_file conf/master/metadata.lua;
         }
 
         # TODO split this into its own file
